Cleaned data from cluster
Added a code block that cleaned the most popular items for Intel.

We realized that one of the clusters in the original had only 19 data points, and through the cleaning process we came to the conclusion that it is worth keeping them in one cluster. this is due to the fact that these movies are so popular that it blows other canidates(aka data points) into another cluster. And through this exploration we have come to the point that it is more important to keep them in there own cluster.